requirements.txt
asyncpg
dateparser
discord.py
humanize
jishaku
psutil
git+https://github.com/Rapptz/discord-ext-menus
